Migrated componenets with context or props to TS
