+++ client/src/components/LoginProvider.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import LoginModalContext from '../contexts/loginModalContext';
+import useLoginModalValues from '../hooks/useLoginModalValues';
+
+type Props = {
+  children: React.ReactNode;
+};
+
+const LoginProvider: React.FC<Props> = (props) => {
+  const values = useLoginModalValues();
+  return (
+    <LoginModalContext.Provider value={values}>
+      {props.children}
+    </LoginModalContext.Provider>
+  );
+};
+
+export default LoginProvider;
